added some catches for saving (or not saving) phylogeny output
RAxML sometimes makes tree files and sometimes doesn't. Instead of
worrying, we just delete stuff if and only if we can find it. This may
introduce some downstream bugs if we're not careful - for instance if
we make a mistake in what the tree_path should be, then trees might be
stacking up (and taking a lot of memory) somewhere odd.

+++ partfinder/analysis.py
@@ -301,8 +301,11 @@ class Analysis(object):
                     if self.save_phylofiles:
                         pass
                     else:
-                        os.remove(stats_path)
-                        os.remove(tree_path)
+                        #we have the ifs because raxml doesn't output a treefile in some cases
+                        if os.path.isfile(stats_path):
+                            os.remove(stats_path)
+                        if os.path.isfile(tree_path):
+                            os.remove(tree_path)
 
                 except self.processor.PhylogenyProgramError:
                     log.warning("Failed loading parse output from %s."
